+++ packages/hotspot/docs/demo/config.js
@@ -1,0 +1,8 @@
+const { model } = require('./generate');
+
+module.exports = {
+  elements: {
+    'hotspot-element': '../..'
+  },
+  models: [model('1', 'hotspot-element')]
+};