+++ packages/hotspot/docs/demo/generate.js
@@ -1,0 +1,25 @@
+exports.model = (id, element) => ({
+  id,
+  element,
+  prompt: 'This is the question prompt',
+  imageUrl: '',
+  shapes: [],
+  multipleCorrect: true,
+  partialScoring: false,
+  dimensions: {
+    height: 0,
+    width: 0
+  },
+  hotspotColor: 'rgba(137, 183, 244, 0.65)',
+  hotspotList: [
+    'rgba(137, 183, 244, 0.65)',
+    'rgba(217, 30, 24, 0.65)',
+    'rgba(254, 241, 96, 0.65)'
+  ],
+  outlineColor: 'blue',
+  outlineList: [
+    'blue',
+    'red',
+    'yellow'
+  ]
+});