+++ packages/inline-dropdown/docs/demo/config.js
@@ -1,0 +1,8 @@
+const { model } = require('./generate');
+
+module.exports = {
+  elements: {
+    'inline-dropdown': '../..'
+  },
+  models: [model('1', 'inline-dropdown')]
+};