+++ packages/ruler/configure/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.8](https://github.com/pie-framework/pie-elements/compare/@pie-element/ruler-configure@2.2.7...@pie-element/ruler-configure@2.2.8) (2019-05-02)
+
+**Note:** Version bump only for package @pie-element/ruler-configure
+
+
+
+
+
+## [2.2.7](https://github.com/pie-framework/pie-elements/compare/@pie-element/ruler-configure@2.2.6...@pie-element/ruler-configure@2.2.7) (2019-05-01)
+
+**Note:** Version bump only for package @pie-element/ruler-configure
+
+
+
+
+
+## [2.2.6](https://github.com/pie-framework/pie-elements/compare/@pie-element/ruler-configure@2.2.5...@pie-element/ruler-configure@2.2.6) (2019-04-26)
+
+
+### Bug Fixes
+
+* Updated with new version of @pie-lib/config-ui. ([e40344b](https://github.com/pie-framework/pie-elements/commit/e40344b))
+
+
+
+
+
+## [2.2.5](https://github.com/pie-framework/pie-elements/compare/@pie-element/ruler-configure@2.2.4...@pie-element/ruler-configure@2.2.5) (2019-04-25)
+
+**Note:** Version bump only for package @pie-element/ruler-configure
+
+
+
+
+
 ## [2.2.4](https://github.com/pie-framework/pie-elements/compare/@pie-element/ruler-configure@2.2.3...@pie-element/ruler-configure@2.2.4) (2019-04-23)